+++ RELEASENOTES.md
@@ -1,1 +1,2 @@
 - Include the migration ID in the default output filename when running `download-logs`
+- When blob storage credentials are provided to the CLI but will not be used for a GHES migration, log a clear warning, not an info message